Update sketch.js
I added a lives counter and a game over text. There is one problem on line 6 i am not completly sure whatt it is but it isn't affecting the game right now. I also gave the movements of our sprite an x and y speed so we can change how fastt it goes in either directiton.